feat: Add JavaScript code to check and print the value of a variable
This commit adds a new JavaScript file, 42.js, which contains code to check the value of a variable and print a corresponding message. The code checks if the variable 'a' is falsy and prints "hello" if it is, otherwise it prints "bye". The variable 'a' is initialized with the value NaN and its boolean value is logged to the console before the check is performed.

The index.html file in section06 is also updated to include the new JavaScript file, 42.js, by uncommenting the script tag.

This commit addresses the need to check and print the value of a variable in JavaScript.