feat: Add JavaScript code to demonstrate variable assignment and object reference
This commit adds JavaScript code to demonstrate variable assignment and object reference. It includes examples of assigning variables, comparing variables, and printing the type and value of variables. The code also showcases how object references work in JavaScript.

The changes include the addition of a new file, "46.js", which contains the code examples. The "index.html" file in the "section06" directory has been updated to include a reference to the new JavaScript file.